% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maury.R
\name{append_ref_alt}
\alias{append_ref_alt}
\title{Append ref, alt alleles to Pileup list}
\usage{
append_ref_alt(apileup, refs, alts)
}
\arguments{
\item{refs}{vector of reference-alleles at different chr-pos}

\item{alts}{vector of alternate-alleles at different chr-pos}

\item{apileup1}{list of lists, each sub-list is for a different chr-pos}
}
\value{
appended list
}
\description{
Append ref, alt alleles to Pileup list
}

