% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maury.R
\name{calc_depths}
\alias{calc_depths}
\title{callback for applyPileups}
\usage{
calc_depths(x)
}
\arguments{
\item{x}{list with elements describing the current pile-up.}
}
\value{
list with seqnames, positions and depths for alleles
}
\description{
callback for applyPileups
}

