% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maury.R
\name{get_best_genotype}
\alias{get_best_genotype}
\title{Return best genotype using MLE}
\usage{
get_best_genotype(ref_rd, alt_rd, seq_error)
}
\arguments{
\item{ref_rd}{readcount for the reference allele}

\item{alt_rd}{readcount for the alternate allele}

\item{seq_error}{sequencing error-rate[0.001]}
}
\value{
genotype 0=R/R, 1=R/A, 2=A/A
}
\description{
Return best genotype using MLE
}

