% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maury.R
\name{get_genotypes}
\alias{get_genotypes}
\title{Compare genotypes of two samples}
\usage{
get_genotypes(apileup1, bams, min_rd, seq_error)
}
\arguments{
\item{apileup1}{list with ref,alt alleles and read-depths}

\item{bams}{vector containing BAMs from both samples}

\item{min_rd}{minimum read-depth to calculate genotypes[20]}
}
\value{
list of both genotypes
}
\description{
Compare genotypes of two samples
}

