% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maury.R
\name{get_metrics}
\alias{get_metrics}
\title{calculate proportion of genotypes that match}
\usage{
get_metrics(gt1, gt2, sample)
}
\arguments{
\item{gt1}{vector of genotypes}

\item{gt2}{vector of genotypes}

\item{sample}{sample-name}
}
\value{
list with four elements -
   sample-name, number of matching GTs, total number of GTs,
   proportion of matching GT's
}
\description{
calculate proportion of genotypes that match
}

