% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maury.R
\name{get_pileup}
\alias{get_pileup}
\title{get pileup at supplied list of sites.}
\usage{
get_pileup(pfiles, vcf_data, min_mq, min_bq)
}
\arguments{
\item{pfiles}{reference to BAM files using Rsamtools::PileupFiles}

\item{vcf_data}{VCF object returned by VariantAnnotation::readVcf}
}
\value{
appended list
}
\description{
get pileup at supplied list of sites.
}

