% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maury.R
\name{maury}
\alias{maury}
\title{Compare two samples' genotypes across a list of given positions}
\usage{
maury()
}
\arguments{
\item{sites}{List of sites to compare genotypes against.}
}
\description{
Compare two samples' genotypes across a list of given positions
}
\examples{
\donttest{
maury(sites = "optimized1.vcf", bam1 = "sample1.bam", bam2 = "sample2.bam")
}
}

